Goal: kalman filter. Done: add sensor_fusion_node.

+++ src/fusion/lidar_process/include/lidar_process/sensor_type.h
@@ -1,11 +1,12 @@
 #ifndef SENSOR_TYPE_H_
 #define SENSOR_TYPE_H_
+#include "Eigen/Dense"
 namespace UstarFusion
 {
     struct LidarClusterType
     {
         Eigen::VectorXd raw_data;
-        long long timestamp;
+        long timestamp;
     };
     
 }
